requirements.txt
appnope==0.1.0
argon2-cffi==19.1.0
atomicwrites==1.3.0
attrs==19.1.0
backcall==0.1.0
bcrypt==3.1.7
beautifulsoup4==4.7.1
bleach==3.0.2
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
decorator==4.3.0
defusedxml==0.5.0
Django==2.1.5
django-cors-headers==3.1.0
djangorestframework==3.10.3
entrypoints==0.2.3
Faker==2.0.0
idna==2.8
ipykernel==5.1.0
ipython==7.2.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.2
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
MarkupSafe==1.1.0
mistune==0.8.4
mock==2.0.0
more-itertools==7.0.0
nbconvert==5.4.0
nbformat==4.4.0
notebook==5.7.4
numpy==1.15.2
pandas==0.23.4
pandocfilters==1.4.2
parso==0.3.1
pbr==5.1.3
pexpect==4.6.0
pickleshare==0.7.5
Pillow==6.1.0
pluggy==0.9.0
prometheus-client==0.5.0
prompt-toolkit==2.0.7
ptyprocess==0.6.0
py==1.8.0
pycparser==2.19
Pygments==2.3.1
PyHamcrest==1.9.0
pytest==4.4.1
python-dateutil==2.7.5
pytz==2018.9
pyzmq==17.1.2
qtconsole==4.4.3
requests==2.21.0
Send2Trash==1.5.0
six==1.12.0
soupsieve==1.7
terminado==0.8.1
testpath==0.4.2
text-unidecode==1.2
tornado==5.1.1
traitlets==4.3.2
urllib3==1.24.2
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.4.2
xmltodict==0.12.0
